Add files via upload
The Snake game is a timeless arcade classic that has engaged players for years with its straightforward yet addictive gameplay. This project showcases the implementation of the Snake game using Python's `turtle` module, which offers a graphical platform for drawing and animation. The game involves controlling a snake that moves across the screen and grows longer each time it consumes a piece of food. The primary goal is to navigate the snake around the screen, avoid collisions with both the game boundaries and itself, and achieve the highest possible score.